requirements/docs.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile '.\requirements\docs.in'
#
alabaster==0.7.12         # via sphinx
attrs==20.3.0             # via -r .\requirements\base.in
babel==2.8.0              # via sphinx
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
colorama==0.4.4           # via sphinx
docutils==0.16            # via sphinx
idna==2.10                # via requests
imagesize==1.2.0          # via sphinx
iso3166==1.0.1            # via -r .\requirements\base.in
jinja2==2.11.2            # via sphinx
markupsafe==1.1.1         # via jinja2
marshmallow-enum==1.5.1   # via -r .\requirements\base.in
marshmallow==3.9.1        # via -r .\requirements\base.in, marshmallow-enum
packaging==20.4           # via sphinx
pygments==2.7.1           # via sphinx
pyparsing==2.4.7          # via packaging
pytz==2020.1              # via babel
requests==2.24.0          # via sphinx
six==1.15.0               # via packaging
snowballstemmer==2.0.0    # via sphinx
sphinx-rtd-theme==0.5.0   # via -r .\requirements\docs.in
sphinx==3.2.1             # via -r .\requirements\docs.in, sphinx-rtd-theme
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
stringcase==1.2.0         # via -r .\requirements\base.in
urllib3==1.25.11          # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
